server/pyproject.toml
[project]
name = "english-conversation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coqui-tts>=0.24.3",
    "fastapi>=0.115.12",
    "faster-whisper>=1.1.0",
    "soundfile>=0.13.1",
    "transformers>=4.51.3",
    "uuid>=1.30",
]
